Move block to being fully server-rendered, and add a deprecation
